Added ImageStream class.
