+++ libssj/src/main/java/hcm/ssj/core/Cons.java
@@ -58,6 +58,7 @@ public class Cons
         DOUBLE,
         BOOL,
         STRING,
+		IMAGE,
         EMPTY //only used for events
     }
